[editor/canvas] Factor out camera state handling. Prep for transition to always using spring state.
