+++ app/src/editor/canvas/components/Camera/index.jsx
@@ -6,6 +6,7 @@ import styles from './Camera.pcss'
 import isEditableElement from '$editor/shared/utils/isEditableElement'
 
 import { useThrottled } from '$shared/hooks/wrapCallback'
+import * as State from './state'
 
 function clamp(value, min, max) {
     return Math.max(min, Math.min(max, value))
@@ -17,368 +18,45 @@ function clamp(value, min, max) {
  */
 export const { cameraControl, noCameraControl } = styles
 
-const defaultFit = {
-    x: 0,
-    y: 0,
-    scale: 1,
-}
-
-function scaleToFit({ bounds, fit, maxScale = 1, padding = 20 } = {}) {
-    const totalPadding = 2 * padding
-    const maxWidth = fit.width - totalPadding
-    const maxHeight = fit.height - totalPadding
-    return Math.min(maxScale, Math.min(maxWidth / bounds.width, maxHeight / bounds.height))
-}
-
-function centerBoundsTo({ bounds, fit, scale }) {
-    // vertically & horizontally center content
-    const offsetY = fit.height / 2
-    const offsetX = fit.width / 2
-
-    return {
-        x: -((bounds.x + (bounds.width / 2)) * scale) + offsetX,
-        y: -((bounds.y + (bounds.height / 2)) * scale) + offsetY,
-    }
-}
-
-function toPrecision(v, precision = 0) {
-    if (!precision) { return v }
-    const p = 10 ** precision
-    return Math.round((v * p)) / p
-}
-
-function fitBoundsTo({ bounds, fit, padding }) {
-    const scale = scaleToFit({
-        bounds,
-        fit,
-        padding,
-    })
-
-    if (!scale) { return defaultFit }
-
-    const { x, y } = centerBoundsTo({
-        bounds,
-        fit,
-        scale,
-    })
-
-    return {
-        x,
-        y,
-        scale,
-    }
-}
-
-function updateScaleState(s, { x, y, scale: newScale }) {
-    const { scale: currentScale } = s
-
-    // adjust for offset created by scale change
-    const ratio = (1 - (newScale / currentScale))
-    const x2 = s.x + ((x - s.x) * ratio)
-    const y2 = s.y + ((y - s.y) * ratio)
-
-    return {
-        ...s,
-        scale: newScale,
-        x: x2,
-        y: y2,
-    }
-}
-const defaultCameraOptions = {
-    scaleFactor: 0.1,
-    scaleLevels: [
-        0.1,
-        0.25,
-        0.5,
-        1.0,
-        1.25,
-        1.50,
-        2.0,
-        3.0,
-    ],
-}
-
-function getParentMatching(el, selector) {
-    if (!el) { return false }
-    if (el.matches(selector)) { return el }
-    return getParentMatching(el.parentElement, selector)
+function useStateCallback(setState, fn) {
+    return useCallback((...args) => (
+        setState((s) => fn(s, ...args))
+    ), [setState, fn])
 }
 
 function useCameraSimpleApi(opts) {
-    const { scaleFactor, scaleLevels } = Object.assign({}, defaultCameraOptions, opts)
-    const minScale = scaleLevels[0]
-    const maxScale = scaleLevels[scaleLevels.length - 1]
-    const [state, setActualState] = useState({
-        scale: 1,
-        x: 0,
-        y: 0,
-    })
-
     const elRef = useRef()
+    const [state, setActualState] = useState(State.createCamera({
+        ...opts,
+        elRef,
+    }))
 
-    const { scale, x, y } = state
+    const stateRef = useRef()
+    stateRef.current = state
 
-    // clamps scale
-    const setState = useCallback((v) => {
-        setActualState((state) => {
-            const nextState = (typeof v === 'function') ? v(state) : v
-            return {
-                ...nextState,
-                scale: clamp(nextState.scale, minScale, maxScale),
-            }
-        })
-    }, [setActualState, minScale, maxScale])
+    const updateScale = useStateCallback(setActualState, State.updateScale)
+    const setPosition = useStateCallback(setActualState, State.setPosition)
+    const pan = useStateCallback(setActualState, State.pan)
+    const setState = useStateCallback(setActualState, State.setState)
+    const zoomIn = useStateCallback(setActualState, State.zoomIn)
+    const zoomOut = useStateCallback(setActualState, State.zoomOut)
+    const setScale = useStateCallback(setActualState, State.setScale)
+    const fitBounds = useStateCallback(setActualState, State.fitBounds)
+    const fitView = useStateCallback(setActualState, State.fitView)
+    const centerBounds = useStateCallback(setActualState, State.centerBounds)
+    const panIntoViewIfNeeded = useStateCallback(setActualState, State.panIntoViewIfNeeded)
 
-    // updates scale of camera, centers scaling on x,y
-    const updateScale = useCallback(({ x, y, delta }) => {
-        setState((s) => {
-            const factor = Math.abs(scaleFactor * (delta / 120))
-            const actualScaleFactor = delta < 0 ? 1 + factor : 1 - factor
-            return updateScaleState(s, {
-                x,
-                y,
-                scale: clamp(s.scale * actualScaleFactor, minScale, maxScale),
-            })
-        })
-    }, [setState, scaleFactor, minScale, maxScale])
-
-    // init needed to reset last values before starting pan
-    const initUpdatePosition = useCallback(({ x, y }) => {
-        setState((s) => ({
-            ...s,
-            lastX: x,
-            lastY: y,
-        }))
-    }, [setState])
-
-    // changes current camera position
-    const setPosition = useCallback(({ x, y }) => {
-        setState((s) => ({
-            ...s,
-            x,
-            y,
-        }))
-    }, [setState])
-
-    // changes current camera offset i.e. pans camera
-    const updatePosition = useCallback(({ x, y }) => {
-        setState(({ lastX = x, lastY = y, ...s }) => ({
-            ...s,
-            lastX: x,
-            lastY: y,
-            x: s.x + (x - lastX),
-            y: s.y + (y - lastY),
-        }))
-    }, [setState])
-
-    // set scale to value, centered on middle of camera el
-    const setScale = useCallback((value) => {
-        setState((s) => {
-            const rect = elRef.current.getBoundingClientRect()
-            return updateScaleState(s, {
-                x: rect.left + (rect.width / 2),
-                y: rect.top + (rect.height / 2),
-                scale: clamp(toPrecision(value, 2), minScale, maxScale),
-            })
-        })
-    }, [setState, minScale, maxScale])
-
-    // zoom in to next largest scale level
-    const zoomIn = useCallback(() => {
-        const currentScale = toPrecision(scale, 2)
-        const nextLevel = scaleLevels.find((v) => v > currentScale) || scale
-        setScale(nextLevel)
-    }, [setScale, scale, scaleLevels])
-
-    // zoom out to next largest scale level
-    const zoomOut = useCallback(() => {
-        const currentScale = toPrecision(scale, 2)
-        const nextLevel = scaleLevels.slice().reverse().find((v) => v < currentScale) || scale
-        setScale(nextLevel)
-    }, [setScale, scale, scaleLevels])
-
-    const pan = useCallback(({ x = 0, y = 0 } = {}) => {
-        setState((s) => ({
-            ...s,
-            x: s.x + x,
-            y: s.y + y,
-        }))
-    }, [setState])
-
-    // fit camera to supplied bounds
-    const fitBounds = useCallback((opts) => {
-        setState((s) => ({
-            ...s,
-            ...fitBoundsTo(opts),
-        }))
-    }, [setState])
-
-    // fit supplied bounds to camera element bounds
-    const fitView = useCallback((opts) => {
-        const { current: cameraEl } = elRef
-        const { width, height } = cameraEl.getBoundingClientRect()
-        return fitBounds({
-            ...opts,
-            fit: {
-                width,
-                height,
-            },
-        })
-    }, [fitBounds, elRef])
-
-    // move bounds into center
-    const centerBounds = useCallback(({ bounds }) => {
-        const { current: cameraEl } = elRef
-        const { width, height } = cameraEl.getBoundingClientRect()
-
-        return setState((s) => ({
-            ...s,
-            ...centerBoundsTo({
-                bounds,
-                fit: {
-                    width,
-                    height,
-                },
-                scale: s.scale,
-            }),
-        }))
-    }, [setState, elRef])
-
-    const areBoundsInView = useCallback(({ bounds, padding = 0 }) => {
-        const { current: cameraEl } = elRef
-        const { width, height } = cameraEl.getBoundingClientRect()
-
-        return (
-            ((bounds.x * scale) + x) >= padding
-            && ((bounds.y * scale) + y) >= padding
-            && (((bounds.x + bounds.width) * scale) + x) < (width - padding)
-            && (((bounds.y + bounds.height) * scale) + y) < (height - padding)
-        )
-    }, [scale, x, y])
-
-    // minimally move bounds into view, if not already in view
-    const panIntoViewIfNeeded = useCallback(({ bounds, padding }) => {
-        const inView = areBoundsInView({
-            bounds,
-            padding,
-        })
-        if (inView) { return }
-
-        const { current: cameraEl } = elRef
-        const { width, height } = cameraEl.getBoundingClientRect()
-
-        const [overLeft, overTop, overRight, overBottom] = [
-            padding - ((bounds.x * scale) + x),
-            padding - ((bounds.y * scale) + y),
-            (((bounds.x + bounds.width) * scale) + x) - (width - padding),
-            (((bounds.y + bounds.height) * scale) + y) - (height - padding),
-        ].map((v) => Math.max(0, v))
-
-        // if too big to fit on screen rescale and center
-        if ((overLeft > 0 && overRight > 0) || (overTop > 0 && overBottom > 0)) {
-            return fitView({
-                bounds,
-                padding,
-            })
-        }
-
-        // prioritise top & left offsets over right & bottom
-        const offsetX = overLeft > 0 ? overLeft : -overRight
-        const offsetY = overTop > 0 ? overTop : -overBottom
-
-        return setState((s) => ({
-            ...s,
-            x: s.x + offsetX,
-            y: s.y + offsetY,
-        }))
-    }, [setState, scale, x, y, areBoundsInView, fitView, elRef])
-
-    // convert bounds in screen/camera coordinates to world coordinates
-    const cameraToWorldBounds = useCallback((bounds) => {
-        const canvasX = (bounds.x - x) / scale
-        const canvasY = (bounds.y - y) / scale
-        return {
-            x: canvasX,
-            y: canvasY,
-            width: bounds.width / scale,
-            height: bounds.height / scale,
-        }
-    }, [x, y, scale])
-
-    // convert point in screen/camera coordinates to world coordinates
-    const cameraToWorldPoint = useCallback((point) => {
-        const { x, y } = cameraToWorldBounds({
-            ...point,
-            width: 1,
-            height: 1,
-        })
-        return {
-            x,
-            y,
-        }
-    }, [cameraToWorldBounds])
-
-    const eventToWorldPoint = useCallback((event) => {
-        const point = {
-            x: 0,
-            y: 0,
-        }
-
-        if (event.target === elRef.current) {
-            // if target is camera el then we can just use event offsets
-            point.x = event.offsetX
-            point.y = event.offsetY
-        } else {
-            const rect = elRef.current.getBoundingClientRect()
-            // otherwise need to calculate offset
-            point.x = event.clientX - rect.left
-            point.y = event.clientY - rect.top
-        }
-
-        const { x, y } = cameraToWorldBounds({
-            ...point,
-            width: 1,
-            height: 1,
-        })
-        return {
-            x,
-            y,
-        }
-    }, [cameraToWorldBounds])
-
-    const shouldIgnoreEvent = useCallback((event) => {
-        const { current: el } = elRef
-
-        // do not ignore if has cameraControl class
-        if (event.target.classList.contains(styles.cameraControl)) { return false }
-        // always ignore if has noCameraControl class
-        if (event.target.classList.contains(styles.noCameraControl)) { return true }
-        // always ignore if target is editable
-        if (isEditableElement(event.target)) { return true }
-
-        // do not ignore if is within parent with cameraControl class
-        const cameraControlEl = getParentMatching(event.target, `.${styles.cameraControl}`)
-        // ignore if is within parent with noCameraControl class
-        const noCameraControlEl = getParentMatching(event.target, `.${styles.noCameraControl}`)
-        if (cameraControlEl && noCameraControlEl) {
-            return cameraControlEl.contains(noCameraControlEl)
-        }
-        if (cameraControlEl) { return false }
-        if (noCameraControlEl) { return false }
-
-        return (
-            // ignore if target not within camera el and isn't body
-            !el.contains(event.target) && event.target !== document.body
-        )
-    }, [elRef])
+    const areBoundsInView = useCallback((...args) => State.areBoundsInView(stateRef.current, ...args), [stateRef])
+    const cameraToWorldBounds = useCallback((...args) => State.cameraToWorldBounds(stateRef.current, ...args), [stateRef])
+    const cameraToWorldPoint = useCallback((...args) => State.cameraToWorldPoint(stateRef.current, ...args), [stateRef])
+    const eventToWorldPoint = useCallback((...args) => State.eventToWorldPoint(stateRef.current, ...args), [stateRef])
+    const shouldIgnoreEvent = useCallback((...args) => State.shouldIgnoreEvent(stateRef.current, ...args), [stateRef])
 
     return useMemo(() => ({
         ...state,
-        elRef,
+        state,
         updateScale,
         setPosition,
-        initUpdatePosition,
-        updatePosition,
         setState,
         zoomIn,
         zoomOut,
@@ -393,15 +71,10 @@ function useCameraSimpleApi(opts) {
         panIntoViewIfNeeded,
         eventToWorldPoint,
         shouldIgnoreEvent,
-        scaleFactor,
-        minScale,
-        maxScale,
     }), [
         state,
         setState,
         updateScale,
-        updatePosition,
-        initUpdatePosition,
         setPosition,
         zoomIn,
         zoomOut,
@@ -416,9 +89,6 @@ function useCameraSimpleApi(opts) {
         panIntoViewIfNeeded,
         eventToWorldPoint,
         shouldIgnoreEvent,
-        scaleFactor,
-        minScale,
-        maxScale,
     ])
 }
 
@@ -451,15 +121,19 @@ function useCameraSpringApi() {
 
     const springRef = useRef()
     const commitSpringState = useCallback(() => {
-        setState({
+        setState((s) => ({
+            ...s,
             x: springRef.current.x.value,
             y: springRef.current.y.value,
             scale: springRef.current.scale.value,
-        })
+        }))
     }, [setState])
 
     const commitDestState = useCallback(() => {
-        setState(destStateRef.current)
+        setState((s) => ({
+            ...s,
+            ...destStateRef.current,
+        }))
     }, [setState])
 
     const [cameraConfig, setCameraConfig] = useState({
@@ -528,7 +202,7 @@ function useCameraSpringApi() {
         const actualScaleFactor = delta < 0 ? 1 + factor : 1 - factor
         const currentScale = destStateRef.current.scale
 
-        destStateRef.current = updateScaleState(destStateRef.current, {
+        destStateRef.current = State.updateScaleState(destStateRef.current, {
             x,
             y,
             scale: clamp(currentScale * actualScaleFactor, minScale, maxScale),
